cover picture and activity date is added to post page

+++ app/Http/Controllers/Admin/Activities/DoActionController.php
@@ -9,6 +9,7 @@ use App\Model\DoActivity ;
 use App\Model\PostGallery ;
 use Illuminate\Support\Facades\Storage;
 use App\Model\SubActivityTitle;
+use \Morilog\Jalali\Jalalian;
 
 class DoActionController extends Controller
 {
@@ -19,7 +20,7 @@ class DoActionController extends Controller
         $Count=null;
         $Posts=null;
         if(count($title)){
-            $Posts =DoActivity::orderBy('created_at','desc')->offset(($Page-1)*10)->limit(10)->get();
+            $Posts =DoActivity::orderBy('apst_accure_date','desc')->offset(($Page-1)*10)->limit(10)->get();
             $Count = DoActivity::count();
         }
         return view('Admin.Activities.Posts.Posts', compact('title','Posts','Count','Page'));
@@ -39,22 +40,36 @@ class DoActionController extends Controller
     public function createPost(Request $request,DoActivity $DoAct){
         $this->validate($request,
                         ['Post_tite' => 'required',
+                        'Post_Date' => 'required',
                         'Post_First_Picture' => 'required',
+                        'Post_Cover_Picture' => 'required',
                         'Post_Description' => 'required',
                         'Catigory' => 'integer',
                         ],
                         ['Post_tite.required' => 'عنوان نبایستی خالی باشد',
-                        'Post_First_Picture.required' => 'فایل تصویر بایستی انتخاب شود',
+                        'Post_Date.required' => 'تاریخ وارد نشده است',
+                        'Post_First_Picture.required' => 'فایل تصویر سربرگ بایستی انتخاب شود',
+                        'Post_Cover_Picture.required' => 'فایل تصویر کاور بایستی انتخاب شود',
                         'Post_Description.required' => 'توضیحات نباید خالی باشد',
                         'Catigory.integer' => 'لطفا یک دسته بندی را انتخاب کنید'
                         ]);
-        $picture = $request->file('Post_First_Picture')->store('picture/post');
-        if($picture){
+        // date validation
+        $dateValidationResut =$this->dateValidation($request->Post_Date);
+        if(!$dateValidationResut["status"]){
+            return redirect()->back()->with('danger','تاریخ اشتباه وارد شده است');
+        }
+        $date = (new Jalalian((int)$dateValidationResut["year"], (int)$dateValidationResut["month"], (int)$dateValidationResut["day"], 0, 0, 0))->toCarbon()->toDateTimeString() ;
+
+        $titlePicture = $request->file('Post_First_Picture')->store('picture/post');
+        $coverPicture = $request->file('Post_Cover_Picture')->store('picture/post');
+        if($titlePicture){
             $DoAct->apst_activities_title_id = $request->Catigory;
             $DoAct->apst_sub_activities_title_id = $request->Sub_Catigory;
             $DoAct->apst_title = $request->Post_tite;
             $DoAct->apst_description = $request->Post_Description;
-            $DoAct->apst_picture_of_title = $picture;
+            $DoAct->apst_picture_of_title = $titlePicture;
+            $DoAct->apst_picture_of_cover = $coverPicture;
+            $DoAct->apst_accure_date = $date;
 
            if($DoAct->save()) {
                 return redirect()->route('Get_Posts')->with('success','پست با موفقیت ثبت شد');
@@ -84,15 +99,28 @@ class DoActionController extends Controller
                         ['Post_tite.required' => 'عنوان نبایستی خالی باشد',
                         'Post_Description.required' => 'توضیحات نباید خالی باشد'
                         ]);
+        // date validation
+        $dateValidationResut =$this->dateValidation($request->Post_Date);
+        if(!$dateValidationResut["status"]){
+            return redirect()->back()->with('danger','تاریخ اشتباه وارد شده است');
+        }
+        $date = (new Jalalian((int)$dateValidationResut["year"], (int)$dateValidationResut["month"], (int)$dateValidationResut["day"], 0, 0, 0))->toCarbon()->toDateTimeString() ;
+
         if(isset($request->Post_First_Picture)){
             Storage::delete($DoAct->apst_picture_of_title);
-            $picture = $request->file('Post_First_Picture')->store('picture/post');
-            $DoAct->apst_picture_of_title = $picture;
+            $titlePicture = $request->file('Post_First_Picture')->store('picture/post');
+            $DoAct->apst_picture_of_title = $titlePicture;
+        }
+        if(isset($request->Post_Cover_Picture)){
+            Storage::delete($DoAct->apst_picture_of_cover);
+            $coverPicture = $request->file('Post_Cover_Picture')->store('picture/post');
+            $DoAct->apst_picture_of_cover = $coverPicture;
         }
         $DoAct->apst_activities_title_id = $request->Catigory;
         $DoAct->apst_sub_activities_title_id = $request->Sub_Catigory;
         $DoAct->apst_title = $request->Post_tite;
         $DoAct->apst_description = $request->Post_Description;
+        $DoAct->apst_accure_date = $date;
         if($DoAct->save()) {
             return redirect()->route('Get_Posts')->with('success','پست با موفقیت ثبت شد');
        }
@@ -142,6 +170,7 @@ class DoActionController extends Controller
             }
         }
         Storage::delete($DeletedRecord->apst_picture_of_title);
+        Storage::delete($DeletedRecord->apst_picture_of_cover);
         PostGallery::where('apic_activities_posts_id',$id)->delete();
         if($DeletedRecord->delete()){
             return back()->with('success','پست با موفقیت حذف شد');
@@ -159,5 +188,17 @@ class DoActionController extends Controller
         $search = 1;
         $title=[1,2];
         return view('Admin.Activities.Posts.Posts',compact('Posts','search','title'));
+    }
+
+    private function dateValidation ($date){
+        $year = $date[0].$date[1].$date[2].$date[3];
+        $month = $date[5].$date[6];
+        $day = $date[8].$date[9];
+
+        if((!is_integer($year) AND $year<1380)) return ["status"=>false];
+        if((!is_integer($month) AND $month>12)) return ["status"=>false];
+        if((!is_integer($day) AND $day>31)) return ["status"=>false];
+        
+        return ["status"=>true, 'year'=> $year , "month" =>  $month , "day" => $day];
     }
 }
